[버그 수정] NULL 오류 테스트 용

+++ Assets/Scripts/SampleClient/Global.cs
@@ -4,7 +4,7 @@ public class Global
 {
     public static byte[] Serialize( object _obj )
     {
-        ushort copySize = ( _obj as UPACKET ).length;
+        int copySize = ( _obj as UPACKET ).length;
         int bufSize = Marshal.SizeOf( _obj );
         byte[] data = new byte[copySize];
         System.IntPtr buffer = Marshal.AllocHGlobal( bufSize + 1 );
